requirements.txt
accelerate
appdirs
loralib
bitsandbytes
black
black[jupyter]
datasets
fire
git+https://github.com/huggingface/peft.git
transformers>=4.28.0
sentencepiece
gradio
tqdm
datasets
sentence_transformers
fuzzywuzzy
python-Levenshtein
pytorch_transformers
tensorboardX
scikit-learn
